Add recovery code hashing and verification functionality
- Updated Makefile to include hash.o in the server build.
- Implemented hash functions for password and recovery code management in src/hash.cpp and src/hash.hpp.
- Modified db_management.cpp to utilize hashed recovery codes for verification and invalidation.
- Updated db_management.hpp to declare new functions for hashed recovery code handling.
- Cleaned up tcp_server.cpp and tcp_server.hpp by removing redundant hashing code and including hash.hpp.
